app/requirements.txt
paho-mqtt==2.1.0
python-dotenv==1.0.1
watchdog==6.0.0
